New - filesystem module
This provides Class['filesystem::base'] to ensure the most essential
file system features are in place and is now included in
Class['base_node'].  This moves the File['/storage'] resource from the
autofs module where it was of questionable fit to this new class where
it fits much better.  This also establishes /storage and /exports on all
hosts with declarations of intent for improved conformity across the
enterprise.