New - rewritten Define[filesystem::mount]
This definition no longer has any remnants of domain-specific details.
As such the mount point must now be fully qualified -- it no longer
contains an implicit prefix.  Nearly all of the parameters from Puppet's
native mount type are exposed here now.  The chief difference with this
definition versus the mount type is that the mount point itself is also
managed.